+++ Updater/ModInfo.cs
